batch_layer/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
yfinance
psycopg2-binary
minio
duckdb
pyarrow
python-dotenv
numpy
boto3